feat(gateway): adressed comments
