+++ gateway-service/src/main/java/org/zowe/apiml/gateway/conformance/OpenApiV2Validator.java
@@ -10,45 +10,46 @@
 
 package org.zowe.apiml.gateway.conformance;
 
-import io.swagger.v3.oas.models.PathItem;
-import io.swagger.v3.parser.OpenAPIV3Parser;
-import io.swagger.v3.parser.core.models.SwaggerParseResult;
+import io.swagger.models.Path;
+import io.swagger.parser.SwaggerParser;
+import io.swagger.parser.util.SwaggerDeserializationResult;
 import org.springframework.http.HttpMethod;
 import org.zowe.apiml.product.gateway.GatewayConfigProperties;
 
 import java.util.*;
 
 
-public class OpenApiV3Parser extends AbstractSwaggerParser {
-    private final SwaggerParseResult swagger;
+public class OpenApiV2Validator extends AbstractSwaggerValidator {
 
+    private final SwaggerDeserializationResult swagger;
 
-    public OpenApiV3Parser(String swaggerDoc, Map<String, String> metadata, GatewayConfigProperties gatewayConfigProperties, String serviceId) {
+    OpenApiV2Validator(String swaggerDoc, Map<String, String> metadata, GatewayConfigProperties gatewayConfigProperties, String serviceId) {
         super(metadata, gatewayConfigProperties, serviceId);
-        swagger = new OpenAPIV3Parser().readContents(swaggerDoc);
+        swagger = new SwaggerParser().readWithInfo(swaggerDoc);
     }
 
     public List<String> getMessages() {
         return swagger.getMessages();
     }
 
-
     public Set<Endpoint> getAllEndpoints() {
         HashSet<Endpoint> result = new HashSet<>();
-        for (Map.Entry<String, PathItem> entry : swagger.getOpenAPI().getPaths().entrySet()) {
-            Set<HttpMethod> methods = getMethod(entry.getValue());
-            String url = generateUrlForEndpoint(entry.getKey());
-            HashMap<String, Set<String>> validResponses = getValidResponses(entry.getValue());
+        for (Map.Entry<String, Path> pathEntry : swagger.getSwagger().getPaths().entrySet()) {
+            Path currentPath = pathEntry.getValue();
+            String currentKey = pathEntry.getKey();
+            Set<HttpMethod> methods = getMethod(currentPath);
+            String url = generateUrlForEndpoint(currentKey);
+            HashMap<String, Set<String>> validResponses = getValidResponses(currentPath);
             Endpoint currentEndpoint = new Endpoint(url, serviceId, methods, validResponses);
             result.add(currentEndpoint);
         }
         return result;
     }
 
-    private HashMap<String, Set<String>> getValidResponses(PathItem value) {
+    private HashMap<String, Set<String>> getValidResponses(Path value) {
         HashMap<String, Set<String>> result = new HashMap<>();
-        for (HttpMethod i : getMethod(value)) {
-            result.put(i.name(), value.readOperationsMap().get(convertSpringHttpToswagger(i)).getResponses().keySet());
+        for (HttpMethod httpMethod : getMethod(value)) {
+            result.put(httpMethod.name(), value.getOperationMap().get(convertSpringHttpToswagger(httpMethod)).getResponses().keySet());
         }
         return result;
     }
@@ -60,31 +61,37 @@ public class OpenApiV3Parser extends AbstractSwaggerParser {
         String version = searchMetadata(metadata, "apiml", "routes", "gatewayUrl");
         String serviceUrl = searchMetadata(metadata, "apiml", "routes", "serviceUrl");
 
+        String baseEndpointPath = swagger.getSwagger().getBasePath();
+
         String endOfUrl;
         if (endpoint.contains("/api/")) {
-            endOfUrl = serviceUrl + endpoint;
+            if (endpoint.indexOf("/api/") > 2) {
+                endOfUrl = endpoint;
+            } else {
+                endOfUrl = serviceUrl + baseEndpointPath + endpoint;
+            }
         } else {
-            endOfUrl = serviceUrl + version + endpoint;
+            endOfUrl = serviceUrl + version + baseEndpointPath + endpoint;
         }
         return baseUrl + endOfUrl.replace("//", "/");
     }
 
-    private Set<HttpMethod> getMethod(PathItem value) {
+    private Set<HttpMethod> getMethod(Path value) {
         Set<HttpMethod> result = new HashSet<>();
-        if (value.getPost() != null) {
-            result.add(HttpMethod.POST);
-        }
         if (value.getGet() != null) {
             result.add(HttpMethod.GET);
-        }
-        if (value.getPatch() != null) {
-            result.add(HttpMethod.PATCH);
         }
         if (value.getHead() != null) {
             result.add(HttpMethod.HEAD);
         }
         if (value.getOptions() != null) {
             result.add(HttpMethod.OPTIONS);
+        }
+        if (value.getPatch() != null) {
+            result.add(HttpMethod.PATCH);
+        }
+        if (value.getPost() != null) {
+            result.add(HttpMethod.POST);
         }
         if (value.getDelete() != null) {
             result.add(HttpMethod.DELETE);
@@ -95,24 +102,23 @@ public class OpenApiV3Parser extends AbstractSwaggerParser {
         return result;
     }
 
-    private PathItem.HttpMethod convertSpringHttpToswagger(HttpMethod input) {
+
+    private io.swagger.models.HttpMethod convertSpringHttpToswagger(HttpMethod input) {
         if (input == HttpMethod.GET) {
-            return PathItem.HttpMethod.GET;
+            return io.swagger.models.HttpMethod.GET;
         } else if (input == HttpMethod.HEAD) {
-            return PathItem.HttpMethod.HEAD;
+            return io.swagger.models.HttpMethod.HEAD;
         } else if (input == HttpMethod.OPTIONS) {
-            return PathItem.HttpMethod.OPTIONS;
+            return io.swagger.models.HttpMethod.OPTIONS;
         } else if (input == HttpMethod.PATCH) {
-            return PathItem.HttpMethod.PATCH;
+            return io.swagger.models.HttpMethod.PATCH;
         } else if (input == HttpMethod.POST) {
-            return PathItem.HttpMethod.POST;
+            return io.swagger.models.HttpMethod.POST;
         } else if (input == HttpMethod.DELETE) {
-            return PathItem.HttpMethod.DELETE;
+            return io.swagger.models.HttpMethod.DELETE;
         } else if (input == HttpMethod.PUT) {
-            return PathItem.HttpMethod.PUT;
+            return io.swagger.models.HttpMethod.PUT;
         }
         return null;
     }
 }
-
-
