+++ gateway-service/src/main/java/org/zowe/apiml/gateway/conformance/ValidatorFactory.java
@@ -13,17 +13,16 @@ package org.zowe.apiml.gateway.conformance;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.experimental.UtilityClass;
 import org.zowe.apiml.product.gateway.GatewayConfigProperties;
 
 import java.util.Map;
 
-
-public class ParserFactory {
-    private ParserFactory() {
-    }
+@UtilityClass
+public class ValidatorFactory {
 
 
-    public static AbstractSwaggerParser parseSwagger(String swaggerDoc, Map<String, String> metadata, GatewayConfigProperties gatewayConfigProperties, String serviceId) {
+    public static AbstractSwaggerValidator parseSwagger(String swaggerDoc, Map<String, String> metadata, GatewayConfigProperties gatewayConfigProperties, String serviceId) {
         JsonNode root;
         try {
             root = new ObjectMapper().readTree(swaggerDoc);
@@ -31,12 +30,13 @@ public class ParserFactory {
             throw new SwaggerParsingException("Could not parse Swagger documentation");
         }
         if (root.findValue("openapi") != null && root.findValue("openapi").asText().split("\\.")[0].equals("3")) {
-            return new OpenApiV3Parser(swaggerDoc, metadata, gatewayConfigProperties, serviceId);
-        } else if (root.findValue("swagger") != null && root.findValue("swagger").asText().equals("2.0")) {
-            return new OpenApiV2Parser(swaggerDoc, metadata, gatewayConfigProperties, serviceId);
-        } else
-            throw new SwaggerParsingException("Swagger documentation is not conformant to either OpenAPI V2 nor V3 - can't " +
-                "find the version (that is cant find field named 'swagger' with value '2.0' or 'openapi' with version starting with '3' )");
+            return new OpenApiV3Validator(swaggerDoc, metadata, gatewayConfigProperties, serviceId);
+        }
+        if (root.findValue("swagger") != null && root.findValue("swagger").asText().equals("2.0")) {
+            return new OpenApiV2Validator(swaggerDoc, metadata, gatewayConfigProperties, serviceId);
+        }
+        throw new SwaggerParsingException("Swagger documentation is not conformant to either OpenAPI V2 nor V3 - cannot " +
+            "find the version (that is cannot find field named 'swagger' with value '2.0' or 'openapi' with version starting with '3' )");
     }
 
 }
